# packages/api-specs/openapi.yaml
openapi: 3.0.0
info:
  title: AI Estimation Platform API
  description: Core REST API for the project estimation engine and developer intelligence models.
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/v1
    description: Local API Gateway

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []

paths:
  # --- ESTIMATION ENDPOINTS ---
  /estimations/generate:
    post:
      summary: Generate a new AI estimate for a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskId:
                  type: string
                teamId:
                  type: string
                sprintId:
                  type: string
      responses:
        '200':
          description: Successfully generated estimate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimationResult'

  /estimations/{id}:
    get:
      summary: Get full details of a specific estimation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Full Estimation object with explanation and factor weights

  /estimations/{id}/override:
    patch:
      summary: Manually override an AI estimate (Requires MANAGER role)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hours:
                  type: number
                reason:
                  type: string
      responses:
        '200':
          description: Estimate successfully overridden

  # --- DEVELOPER PROFILE ENDPOINTS ---
  /developers/{id}/profile:
    get:
      summary: Retrieve a developer's multi-dimensional behavioral profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: DeveloperProfile object

  # --- SIMULATION ENGINE ENDPOINTS ---
  /simulations/what-if:
    post:
      summary: Run a scenario simulation on project timelines
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectId:
                  type: string
                changes:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: SimulationResult containing timeline, cost, and risk deltas

components:
  schemas:
    EstimationResult:
      type: object
      properties:
        optimisticHours:
          type: number
        expectedHours:
          type: number
        pessimisticHours:
          type: number
        confidenceScore:
          type: number